get rid of warnings
As of Nutils 7, function arrays can be manipulated directly using numpy
operations such as numpy.sqrt and numpy.log2. With that, the equivalent methods
in the function module have lost their purpose and are due to be removed in
Nutils 9. This patch substitutes function.sqrt and function.log2 by their numpy
equivalents to remove the deprecation warnings a be forward compatible with
Nutils 9.