simplify get_elements_to_be_refined algorithm
This patch simplifies the get_elements_to_be_refined method by switching from a
refinement + connectivity based approach to a direct lookup of interface
transformations.

The main advantage of the new approach is brevity, but it also positions the
algorithm better for future Nutils developments, which will likely gravitate
towards operations on entire sequences rather than dealing with individual
transformations chains. In this future, the 'for trans in' loop will be
rewritten as a single lookup operation from the entire interface transformation
sequence into the topology deformation sequence, which is potentially much more
efficient as it can exploit sequence-level structure that currently needs be be
rediscovered for every transformation at a time.